more level building

+++ november/gameLevelHandler.js
@@ -81,13 +81,15 @@ class Level {
     this.dialogCondition = [];
     this.dialogs = [];
   }
-  setSpritesToAttack() {
+  setSpritesToAttack(deleteDialogs=true) {
     for (let i = 0; i < this.npcs.length; i++) {
       this.npcs[i].autochange = true;
       this.npcs[i].attack = true;
       this.npcs[i].aggressive = true;
       this.npcs[i].doRandomWalk = true;
-      this.deleteDialogs();
+      if (deleteDialogs == true) {
+        this.deleteDialogs();
+      }
     }
     return true;
   }
